--- neulib/DAHosts.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace neulib
-{
-    [Serializable]
-    public class DAHostsReqMsg : MsgBase
-    {
-    }
-
-    [Serializable]
-    public class DAHostsResMsg : MsgBase
-    {
-        public List<string> Hosts;
-    }
-}
