--- neulib/Exit.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace neulib
-{
-    [Serializable]
-    public class ExitReqMsg : MsgBase
-    {
-    }
-
-    [Serializable]
-    public class ExitResMsg : MsgBase
-    {
-    }
-}
