--- neulib/PingPong.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace neulib
-{
-    [Serializable]
-    public class PingMsg : MsgBase
-    {
-    }
-
-    [Serializable]
-    public class PongMsg : MsgBase
-    {
-    }
-}
