+++ neuopc.sln
@@ -5,13 +5,11 @@ VisualStudioVersion = 17.7.34018.315
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "neuopc", "neuopc\neuopc.csproj", "{495EB8D0-766F-494F-B60B-2E46A9DFB7A4}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "neuservice", "neuservice\neuservice.csproj", "{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "neulib", "neulib\neulib.csproj", "{49F6BB27-3EAE-4AD5-B081-CE896A33D4DF}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "neuclient", "neuclient\neuclient.csproj", "{8E067467-89FE-45C5-A9FE-A2EAEBD96219}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "neuserver", "neuserver\neuserver.csproj", "{52AC8A1E-B09D-4977-9D67-CE232CFDBF19}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "neuserver", "neuserver\neuserver.csproj", "{52AC8A1E-B09D-4977-9D67-CE232CFDBF19}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,14 +27,6 @@ Global
 		{495EB8D0-766F-494F-B60B-2E46A9DFB7A4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{495EB8D0-766F-494F-B60B-2E46A9DFB7A4}.Release|x86.ActiveCfg = Release|x86
 		{495EB8D0-766F-494F-B60B-2E46A9DFB7A4}.Release|x86.Build.0 = Release|x86
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Debug|x86.ActiveCfg = Debug|x86
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Debug|x86.Build.0 = Debug|x86
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Release|x86.ActiveCfg = Release|x86
-		{E0EA2C0E-57CD-4AC4-A491-5C4E0EB9E405}.Release|x86.Build.0 = Release|x86
 		{49F6BB27-3EAE-4AD5-B081-CE896A33D4DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{49F6BB27-3EAE-4AD5-B081-CE896A33D4DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{49F6BB27-3EAE-4AD5-B081-CE896A33D4DF}.Debug|x86.ActiveCfg = Debug|Any CPU
